pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teamworksams"
version = "0.1.0"
description = "A Python library for interacting with the Teamworks AMS API"
readme = "README.md"
authors = [{name = "Brandon Yach", email = "yachb35@gmail.com"}]
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.2.3",
    "requests",
    "requests_toolbelt",
    "python-dotenv",
    "tqdm",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
test = ["pytest>=7.0", "pytest-cov>=4.0", "pytest-vcr", "responses>=0.23"]

[project.urls]
Homepage = "https://github.com/brandonyach/teamworksams"
Documentation = "https://brandonyach.github.io/teamworksams"